refactor: move legacy resolver to webpack plugin package
