+++ packages/experimental-loader/package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "@stylable/core": "^5.16.0",
     "@stylable/runtime": "^5.16.0",
+    "@stylable/webpack-plugin": "^5.16.0",
     "css-loader": "^6.8.1",
     "decache": "^4.6.2",
     "loader-utils": "^3.2.1"
