+++ packages/webpack-plugin/src/legacy-module-resolver.ts
@@ -3,8 +3,7 @@ import nodeModule from 'module';
 // importing the factory directly, as we feed it our own fs, and don't want graceful-fs to be implicitly imported
 // this allows @stylable/core to be bundled for browser usage without special custom configuration
 import ResolverFactory from 'enhanced-resolve/lib/ResolverFactory.js';
-import type { ModuleResolver } from './types';
-import type { MinimalFS } from './cached-process-file';
+import type { MinimalFS } from '@stylable/core';
 
 function bundleSafeRequireExtensions(): string[] {
     let extensions: string[];
@@ -22,7 +21,7 @@ function bundleSafeRequireExtensions(): string[] {
 
 const resolverContext = {};
 
-export function createLegacyResolver(fileSystem: MinimalFS, resolveOptions: any): ModuleResolver {
+export function createLegacyResolver(fileSystem: MinimalFS, resolveOptions: any) {
     const extensions =
         resolveOptions.extensions && resolveOptions.extensions.length
             ? resolveOptions.extensions
@@ -35,7 +34,7 @@ export function createLegacyResolver(fileSystem: MinimalFS, resolveOptions: any)
         fileSystem,
     });
 
-    return (directoryPath, request): string => {
+    return (directoryPath: string, request: string): string => {
         const res = eResolver.resolveSync(resolverContext, directoryPath, request);
         if (res === false) {
             throw new Error(
