+++ packages/webpack-plugin/src/plugin.ts
@@ -1,4 +1,4 @@
-import { MinimalFS, Stylable, StylableConfig, createLegacyResolver } from '@stylable/core';
+import { MinimalFS, Stylable, StylableConfig } from '@stylable/core';
 import {
     OptimizeConfig,
     DiagnosticsMode,
@@ -49,6 +49,7 @@ import type {
 import { parse } from 'postcss';
 import { getWebpackEntities, StylableWebpackEntities } from './webpack-entities';
 import { resolveConfig as resolveStcConfig, STCBuilder } from '@stylable/cli';
+import { createLegacyResolver } from './legacy-module-resolver';
 
 type OptimizeOptions = OptimizeConfig & {
     minify?: boolean;
