+++ packages/webpack-plugin/test/e2e/projects/stylable-config-resolver/stylable.config.js
@@ -1,14 +1,10 @@
 //@ts-check
 const { join } = require('path');
-const { createLegacyResolver } = require('@stylable/core');
+const { createLegacyResolver } = require('@stylable/webpack-plugin');
 const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
 
 module.exports = {
     defaultConfig(fs) {
-        /////////////////////////////////////////
-        /////////////////////////////////////////
-        /////////////////////////////////////////
-        /////////////////////////////////////////
         return {
             resolveModule: createLegacyResolver(fs, {
                 alias: {
